Add methods to authorise and validate a type

+++ app/api/mutations/vote_mutations.rb
@@ -8,6 +8,8 @@ module VoteMutations
     resolve ->(_obj, inputs, ctx) {
       votable = RelaySchema.object_from_id(inputs[:votable_id], ctx)
       user = ctx[:current_user]
+
+      authorize(user, Vote, :create)
       votable.votes.create(user: user)
 
       { votable.class.to_s.downcase.to_sym => votable.reload }
@@ -23,7 +25,9 @@ module VoteMutations
     resolve ->(_obj, inputs, ctx) {
       votable = RelaySchema.object_from_id(inputs[:votable_id], ctx)
       user = ctx[:current_user]
+
       vote = user.votes.where(votable: votable).first
+      authorize(user, vote, :destroy)
 
       vote.destroy
